Added Logger Implemented Exception Handling

+++ src/main/java/com/ems/employee_management_system/service/EmployeeService.java
@@ -2,6 +2,8 @@ package com.ems.employee_management_system.service;
 
 import com.ems.employee_management_system.entity.Department;
 import com.ems.employee_management_system.entity.Employee;
+import com.ems.employee_management_system.exception.EmployeeNotFoundException;
+import com.ems.employee_management_system.exception.InvalidDepartmentException;
 import com.ems.employee_management_system.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,7 +22,12 @@ public class EmployeeService {
 
     //Adding a single employee to DB
     public Employee addEmployee(String name , int deptId , double salary) {
-        Department dept = Department.fromId(deptId);
+        Department dept;
+        try {
+            dept = Department.fromId(deptId);
+        } catch(IllegalArgumentException e) {
+            throw new InvalidDepartmentException(deptId);
+        }
         Employee emp = new Employee(name,dept,salary);
         return repository.save(emp);
     }
@@ -32,7 +39,7 @@ public class EmployeeService {
 
     //Retrieving employee by ID
     public Employee getEmployeeById(int id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("Employee not found"));
+        return repository.findById(id).orElseThrow(() -> new EmployeeNotFoundException(id));
     }
 
     //Retrieving employee by name
@@ -47,14 +54,23 @@ public class EmployeeService {
 
     //Deleting employee records by ID
     public void deleteEmployee(int id) {
+        if(!repository.existsById(id)) {
+            throw new EmployeeNotFoundException(id);
+        }
         repository.deleteById(id);
     }
 
     //Updating employee details
     public Employee updateEmployee(int id , String name , int deptId , double salary) {
+        Department dept;
+        try {
+            dept = Department.fromId(deptId);
+        } catch(IllegalArgumentException e) {
+            throw new InvalidDepartmentException(deptId);
+        }
         Employee emp = getEmployeeById(id);
         emp.setName(name);
-        emp.setDept(Department.fromId(deptId));
+        emp.setDept(dept);
         emp.setSalary(salary);
         repository.save(emp);
         return emp;
